Squashed 'android/src/main/java/io/agora/rtc/base/' changes from 21c7655..987de50
987de50 feat: support 3.5.2
4ebf77c feat: support 3.5.1 & plugin
27c788c chore: mapToColor
0278629 feat: support 3.5.0.2
45dcd16 chore: add API which is missing
58ca4b0 fix: AudioRecordingConfiguration bug
4a32e12 chore: prepare to migrate to iris
276c9fe chore: support 3.4.5
783aecf chore: add ERR_ALREADY_IN_RECORDING
3aed6ce fix: startAudioRecording bug
510b405 chore: optimize
4f801b9 chore: optimize
ab71f52 chore: optimize startAudioRecording
07f81c7 chore: optimize
5a20e13 style: optimize
858b58f fix(uid compatible): int to uint
8d14d2d feat(upgrade): 3.4.1
be5efea Merge pull request #4 from AgoraLibrary/dev/3.3.0
382f858 Merge branch 'master' into dev/3.3.0
8671535 Merge pull request #3 from AgoraLibrary/dev/3.2.0
feabda3 Merge branch 'master' into dev/3.2.0
9e57891 Merge pull request #2 from AgoraLibrary/dev/3.1.0
df303ee fix: merge some bug fix
cd7a2cd feat: support 3.3.1 for Android
551ebfe fix: annotation build warning
9c48594 feat: support 3.3.0 for Android
de52ca6 Merge branch 'master' into dev/3.2.0
ddeaefb Merge branch 'master' into dev/3.2.0
2fef3ae feat: add `setClientRole(role: ClientRole, options?: ClientRoleOptions): Promise<void>`
fcb57cc feat: upgrade to 3.2.0
REVERT: 21c7655 feat: support 3.0.+

git-subtree-dir: android/src/main/java/io/agora/rtc/base
git-subtree-split: 987de50b775615157512823211fd86be2c344493

+++ MediaObserver.kt
@@ -31,8 +31,10 @@ class MediaObserver(
   }
 
   override fun onMetadataReceived(buffer: ByteArray, uid: Int, timeStampMs: Long) {
-    emit(hashMapOf(
-      "data" to arrayListOf(String(buffer), uid, timeStampMs)
-    ))
+    emit(
+      hashMapOf(
+        "data" to arrayListOf(String(buffer), uid.toUInt().toLong(), timeStampMs)
+      )
+    )
   }
 }
